requirements_test.txt
coverage==5.3
django-appconf
flake8>=3.8.3
mock>=3.0.5
pathvalidate==2.3.0
pytest-cov
pytest-django
pytest-logger
pytest-pythonpath
tox>=3.20.0
tuspy
